Move StringMixin to computations.common
