+++ pandas/core/computation/ops.py
@@ -12,7 +12,6 @@ from pandas._libs.tslibs import Timestamp
 
 from pandas.core.dtypes.common import is_list_like, is_scalar
 
-from pandas.core.base import StringMixin
 import pandas.core.common as com
 from pandas.core.computation.common import _ensure_decoded, _result_type_many
 from pandas.core.computation.scope import _DEFAULT_GLOBALS
@@ -63,7 +62,7 @@ class UndefinedVariableError(NameError):
         super().__init__(msg.format(name))
 
 
-class Term(StringMixin):
+class Term:
     def __new__(cls, name, env, side=None, encoding=None):
         klass = Constant if not isinstance(name, str) else cls
         supr_new = super(Term, klass).__new__
@@ -82,7 +81,7 @@ class Term(StringMixin):
     def local_name(self):
         return self.name.replace(_LOCAL_TAG, "")
 
-    def __str__(self):
+    def __repr__(self):
         return pprint_thing(self.name)
 
     def __call__(self, *args, **kwargs):
@@ -182,7 +181,7 @@ class Constant(Term):
     def name(self):
         return self.value
 
-    def __str__(self):
+    def __repr__(self):
         # in python 2 str() of float
         # can truncate shorter than repr()
         return repr(self.name)
@@ -191,7 +190,7 @@ class Constant(Term):
 _bool_op_map = {"not": "~", "and": "&", "or": "|"}
 
 
-class Op(StringMixin):
+class Op:
 
     """Hold an operator of arbitrary arity
     """
@@ -204,7 +203,7 @@ class Op(StringMixin):
     def __iter__(self):
         return iter(self.operands)
 
-    def __str__(self):
+    def __repr__(self):
         """Print a generic n-ary operator and its operands using infix
         notation"""
         # recurse over the operands
@@ -537,7 +536,7 @@ class UnaryOp(Op):
         operand = self.operand(env)
         return self.func(operand)
 
-    def __str__(self):
+    def __repr__(self):
         return pprint_thing("{0}({1})".format(self.op, self.operand))
 
     @property
@@ -562,7 +561,7 @@ class MathCall(Op):
         with np.errstate(all="ignore"):
             return self.func.func(*operands)
 
-    def __str__(self):
+    def __repr__(self):
         operands = map(str, self.operands)
         return pprint_thing("{0}({1})".format(self.op, ",".join(operands)))
