+++ pandas/core/computation/scope.py
@@ -15,8 +15,8 @@ import numpy as np
 from pandas._libs.tslibs import Timestamp
 from pandas.compat.chainmap import DeepChainMap
 
-from pandas.core.base import StringMixin
 import pandas.core.computation as compu
+from pandas.core.computation.common import StringMixin
 
 
 def _ensure_scope(
@@ -141,7 +141,7 @@ class Scope(StringMixin):
         self.resolvers = DeepChainMap(*resolvers)
         self.temps = {}
 
-    def __str__(self):
+    def __repr__(self):
         scope_keys = _get_pretty_string(list(self.scope.keys()))
         res_keys = _get_pretty_string(list(self.resolvers.keys()))
         unicode_str = "{name}(scope={scope_keys}, resolvers={res_keys})"
